Update from deprecated jung-kurt pdf library to phpdave11.
I'm trying to clean up some archived dependencies from my tree, and
I landed here! This repo is getting included through some transitive deps of
mine, and is still using jung-kurt/gofpdf in an example. phpdave11's fork seems
to be the most active.

I also updated the go version in Travis to 1.15. I hope this is OK! 1.11 isn't supported
anymore, and this doesn't build with it.